+++ stage0/src/compiler_common/mod.rs
@@ -14,6 +14,7 @@ use classfile_view::view::method_view::MethodView;
 use gc_memory_layout_common::layout::{FRAME_HEADER_END_OFFSET, NativeStackframeMemoryLayout};
 use gc_memory_layout_common::memory_regions::{AllocatedTypeID, RegionHeader};
 use inheritance_tree::ClassID;
+use jvmti_jni_bindings::jint;
 use method_table::interface_table::InterfaceID;
 use method_table::MethodTable;
 use runtime_class_stuff::method_numbers::MethodNumber;
@@ -179,7 +180,7 @@ pub trait MethodResolver<'gc> {
     fn lookup_interface_method_number(&self, interface: CPDType, method_shape: MethodShape) -> Option<MethodNumber>;
     fn lookup_special(&self, on: &CPDType, name: MethodName, desc: CMethodDescriptor) -> Option<(MethodId, bool)>;
     fn lookup_type_inited_initing(&self, cpdtype: &CPDType) -> Option<(Arc<RuntimeClass<'gc>>, LoaderName)>;
-    fn allocated_object_type_id(&self, rc: Arc<RuntimeClass<'gc>>, loader: LoaderName, arr_len: Option<usize>) -> AllocatedTypeID;
+    fn allocated_object_type_id(&self, rc: Arc<RuntimeClass<'gc>>, loader: LoaderName, arr_len: Option<jint>) -> AllocatedTypeID;
     fn allocated_object_region_header_pointer(&self, id: AllocatedTypeID) -> *const AtomicPtr<RegionHeader>;
     fn lookup_method_layout(&self, method_id: usize) -> YetAnotherLayoutImpl;
     fn lookup_native_method_layout(&self, method_id: usize) -> NativeStackframeMemoryLayout;
